requirements.txt
colorthief
firebase_admin
flask
flask-cors
Pillow
gunicorn
yagmail